make separate search file
